SystemUI: Prevent NPE in Dependency.get() in BrightnessController
When com.android.systemui process died for some reason, the process will
be restarted and will schedule to restart services, such as
SystemUIService and KeyguardService, if start activity BrightnessDialog
before these services, then NPE will happen, the reason is that this
activity will use Dependency which is initalized in these services.

Init Dependency in BrightnessController before using it.

Change-Id: I83f490bf3380a79991c9219ee0238e77a164bf7c
CRs-Fixed: 2083957
Signed-off-by: mydongistiny <jaysonedson@gmail.com>
Signed-off-by: NurKeinNeid <mralexman3000@gmail.com>